// sdk-1.x/src/interfaces.ts
export interface PublicKeyModel {
    Curvname: string;
    X: string;
    Y: string;
}

/**
 * Model Interface - Private key
 */
export interface PrivateKeyModel extends PublicKeyModel {
    D: string;
}

/**
 * Model Interface - Xuper account
 */
export interface AccountModel {
    /**
     * Account address
     */
    address: string;

    /**
     * Account mnemonic
     */
    mnemonic?: string;

    /**
     * Account public key model
     */
    publicKey: PublicKeyModel;

    /**
     * Account private key model
     */
    privateKey: PrivateKeyModel;
}
// ...
export interface TXInput {
    refTxid: string;
    refOffset: number;
    fromAddr: string;
    amount: string;
    frozenHeight?: number;
}

/**
 * @hidden
 * Model Interface - transaction output
 */
export interface TXOutput {
    amount: string;
    toAddr: string;
}

/**
 * @hidden
 * Model Interface - UTXO
 */
export interface UTXO {
    amount: string;
    toAddr: string;
    refOffset: number;
    refTxid: string;
}

/**
 * @hidden
 * Model Interface - signinfo
 */
export interface SignInfoModel {
    PublicKey: string;
    Sign: string;
}

/**
 * @hidden
 * Model Interface - transaction infomation
 */
export interface TransactionInfomation {
    amount: string | number | BN;
    fee: string | number | BN;
    to: string;
    desc?: string;
}

/**
 * Model Interface - Transaction
 */
export interface Transaction {
    /**
     * SDK version
     */
    version: number;

    /**
     * Timestamp
     */
    timestamp: number;

    coinbase: boolean;
    autogen: boolean;

    /**
     * Transaction description
     */
    desc: string;

    /**
     * Inputs
     */
    txInputs: TXInput[];

    /**
     * Outputs
     */
    txOutputs: any[];

    /**
     * Transaction initiator
     */
    initiator: string;

    /**
     * initiator sign
     */
    initiatorSigns?: SignInfoModel[];

    /**
     * Nonce
     */
    nonce: string;

    /**
     * auth
     */
    authRequire: string[];

    /**
     * auth sign
     */
    authRequireSigns?: any[];

    /**
     * Transaction ID
     */
    txid?: string;

    txInputsExt?: any[];

    txOutputsExt?: any[];

    contractRequests?: any[];
}